requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
Django==5.1.4
django-apscheduler==0.7.0
django-extensions==3.2.3
django-solo==2.4.0
django-unfold==0.43.0
Faker==33.1.0
graphviz==0.20.3
numpy==2.2.1
pandas==2.2.3
psycopg==3.2.3
pyaml==25.1.0
pydicom==3.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.3
tzdata==2024.2
tzlocal==5.2
wheel==0.44.0
python-dotenv == 1.0.1
django-allauth-ui
django-widget-tweaks
django-allauth
slippers
django-admin-tools
django-unfold
requests
whitenoise
django-log-viewer